feat: big updates

+++ pkg/cmhtsl2591/tsl2591.go
@@ -10,8 +10,8 @@ import (
 	tsl2591 "github.com/JenswBE/golang-tsl2591"
 )
 
-func LightChannel(ctx context.Context, dev *tsl2591.TSL2591, interval time.Duration) (<-chan uint64, func() error) {
-	c := make(chan uint64, 1)
+func LightChannel(ctx context.Context, dev *tsl2591.TSL2591, interval time.Duration) (<-chan float64, func() error) {
+	c := make(chan float64, 1)
 	ctx, cancelFunc := context.WithCancel(ctx)
 	return c, func() error {
 		defer cancelFunc()
@@ -27,7 +27,7 @@ func LightChannel(ctx context.Context, dev *tsl2591.TSL2591, interval time.Durat
 					return fmt.Errorf("tsl2591: %w", err)
 				}
 				slog.Debug("publishing reading", "ir", ir, "module", "tsl2591")
-				c <- uint64(ir)
+				c <- float64(ir)
 			}
 		}
 	}
